Bug 1945748 - Expose caret blink count / rate to script. r=Gijs
For a variety of reasons (see bug 1933080 and bug 1508388), the new tab
page needs to fake a caret. This will allow them to fake it more
reliably.

I thought of just exposing this to CSS via an env() var, but turning the
negatives into something usable generically seems a bit hard (e.g. one
should be calc(inf) or so, and the other should somehow not animate at
all, which is not trivial to express on a single value.

Differential Revision: https://phabricator.services.mozilla.com/D236676

UltraBlame original commit: 1a152fa55556a1605e99c2324a0e524c229e5c6d