Bug 1944908 - Improve logging in RemoteSettingsExperimentLoader r=emcminn
The id and slug fields now always match in generated recipes.
Previously, the id field was always a UUID. However, in production, the
id field is always equal to the slug field. Now our generated
experiments in tests more closely match real experiments.

Additionally, loggging statements in RemoteSettingsExperimentLoader have
been updated to use the slug instead of the id, as that is preferred as
the source of truth of what an experiment is named.

Finally, ExperimentFakes.recipe now generates recipes with targeting =
"true" instead instead of undefined. This is functionally equivalent, as
a missing targeting field is interpreted as targeting = "true", but we
also log that seperately special occurance. This reduces unnecessary
logging.

Differential Revision: https://phabricator.services.mozilla.com/D236245

UltraBlame original commit: 47039dc3abd1f09212b7ac65f0434318fcb11adc