Bug 1908583: Skip reftest selection/1478604.html on Android emulator for frequent failures. r=hiro
This patch replaces the stale code-comment with an up-to-date one for the skip
annotation that I'm adding here.

The stale comment was documenting a fuzzy annotation that was removed long ago,
which was specific to non-webrender builds (which aren't possible / supported
at this point).

Differential Revision: https://phabricator.services.mozilla.com/D236725

UltraBlame original commit: 802dcc7fb6d58a355327452e0f73fb213c1e37b0