Bug 1946016 - Don't be explicit about window size when detaching tabs on Linux. r=dao,tabbrowser-reviewers
Let Gecko choose a suitable size. This seems to result in the desired
behavior (same as Ctrl+N). Tested GNOME and Kwin, X11 and Wayland.

The reason for the behavior change here is because `outerWidth` before
the regressing patch returned the inner, not the outer width. Combined
with the pre-existing behavior that decorations are added asynchronously
by the window manager, it just happened to conveniently cancel out that
bug, which is now fixed.

It's unclear we need this at all on macOS fwiw. IIRC Cmd+N does also the
same thing as Windows, so maybe we can remove this block altogether.

Separate patch tho, as I don't have a macOS build ready to test it this
very moment, but I think it should just work.

Differential Revision: https://phabricator.services.mozilla.com/D236867

UltraBlame original commit: dbbe4e0b455e56ba0412da0435d494192951efd0