Bug 1944365 - Part 3: Store nsSimpleURI spec in contiguous memory, r=necko-reviewers,valentin
This patch adjusts nsSimpleURI to be stored as a single `mSpec` member,
with cached indices of the various separators alongside. This improves
the performance of calling `GetSpec()` on these URIs, as the caller will
be able to share the underlying string buffer.

This moves nsSimpleURI's implementation to be a bit more in-line with
nsStandardURL, hopefully reducing performance surprises, especially
around large data URIs.

nsDataChannel is updated for the new perf characteristics in part 4.

Differential Revision: https://phabricator.services.mozilla.com/D236274

UltraBlame original commit: 015322a112ce2099639d7897ab689da777e634d7