Bug 1945894 - Don't bother tracking opaque region on popups. r=win-reviewers,rkraesig
So the issue here is a mix of things:

 1. Our opaque region tracking goes wrong (that's expected-ish, because
    places popups are scrollable, and our opaque region tracking is a
    bit suboptimal, see bug 1933952). So every time the "drop" marker
    moves we think everything _but_ the marker is transparent.

 2. We don't track areas of the window we've already cleared (we could
    improve that on the widget side), so we clear all the transparent
    region (which is not transparent because of the above!) every time
    the opaque region changes. We could fix that, but it's tricky to do
    across resizes and a bigger change. Probably worth it tho.

 3. The places code is doing a lot of relatively slow stuff on this case
    which invalidates the popup too much.

The simplest fix for now is to do what the old code did and just assume
transparent popups are fully translucent all the time. That doesn't
change much in practice (because of bug 1933952), but it does fix this
issue and it's a straight-forward change.

I could look into (2) separately if you think it's worth it.

Differential Revision: https://phabricator.services.mozilla.com/D236847

UltraBlame original commit: 3fd134c1e63f0dd8b9c337200316847247e20a44