Bug 1946357 - Prefer Microsoft JhengHei as sans-serif font for zh-TW locale on Windows for all release channels. r=dshin,TYLin
We've been shipping this configuration on Nightly/early-Beta since bug 1394709
(7 years ago), but temporarily reverted it for late-Beta/Release in bug 1517937
due to issues with the JhengHei font and some users' preference for the old
PMingLiu face.

But as reported in this bug, PMingLiu really isn't an appropriate mapping for
'sans-serif', and as :petercpg suggests in [1], it seems like time to ship
this change.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1946357#c8

Differential Revision: https://phabricator.services.mozilla.com/D237574

UltraBlame original commit: 5738d17b06cd0b9f82c8f7aead434f2768badafe