Bug 1946445 - [devtools] Fix parseDeclarationsInternal for nested declarations. r=devtools-reviewers,jdescottes.
Stop collecting declarations as soon as we detect a nested rule.
Any declaration that come after the nested rules are now put in a
separate CSSNestedDeclaration rule.
We still need to handle the nesting logic for commented declarations,
as highlighted by a test case in test_parseDeclarations.js.
Other test cases there for nested rules are updated, and we add
a CSSNestedDeclaration in a Rules view mochitest to make sure we
don't get the declaration in the "top level" rule anymore.

Differential Revision: https://phabricator.services.mozilla.com/D237520

UltraBlame original commit: a08727399a3753b89c1270329ebd0f0499c3740d