Bug 1936020: Part 6 - Remove flavors from content analysis cache r=dlp-reviewers,gstoll
Removed flavors from cache since they were not being considered in verdicts.
The cache is all or nothing so it added nothing to the seqno comparison.
This does not include the changes to the non-test callers, which are in another
patch.

Differential Revision: https://phabricator.services.mozilla.com/D236629

UltraBlame original commit: 041abc3c071aa803a5ef85ce8e1f3cf5fe0dd15a