Bug 1946988 - Ensure Translations tests wait for model records changed r=translations-reviewers,gregtatum
This patch ensures that when the useLexicalShortlist pref changes
the relevant Translations tests wait for the model records to be
modified, instead of only waiting for the pref to change. This will
likely fix the race condition causing the intermittent test failure
associated with this bug.

Differential Revision: https://phabricator.services.mozilla.com/D237928

UltraBlame original commit: f1ed11b4d153023976b2a16cb4baa4c49cc70776