Bug 1944669 - Suppress ProcessWatcher logspam introduced in earlier rewrite. r=ipc-reviewers,mccr8
The case in question is where the process exits before the timeout
expires and it's a forkserver child.  Previously (before bug 1658072)
there wouldn't be a log message in this case, but otherwise the effect
was the same: if a process with the given pid was still running, it
would be sent `SIGKILL`.

Differential Revision: https://phabricator.services.mozilla.com/D237857

UltraBlame original commit: 3e357dbb1efcec6eae8f167c314d4180858ea1de