Bug 1948283 - Less bounds recomputation. r=stransky
I misunderstood the size_allocate event setup. We only get it on the
container, so we can just simplify a bit and go back to not recomputing
when the size hasn't changed.

Also simplify configure to put in the name that we only listen to
mShell's configure.

Differential Revision: https://phabricator.services.mozilla.com/D238026

UltraBlame original commit: fe5e1ebb9852cdf520c2bfd51ab71d46c1859a6d