Bug 1946442 - [devtools] Fix RuleRewriter when dealing with rules containing nested rules. r=devtools-reviewers,jdescottes.
Create a dummy rule with the content of the rule being modified so we can
check if it has a nested rule. In such case, we can then get the location
of the nested rule so we can insert the added text in the correct position.

A few unit test cases are added to make sure the RuleRewriter does handle
nested rules properly.

Differential Revision: https://phabricator.services.mozilla.com/D237734

UltraBlame original commit: 850b58556292ebe1fb631193c190ceb7481f1acd