Bug 1947767 part 5 - Make Watchtower::watchPropertyValueChange infallible. r=mgaudet
The only reason for this to be fallible is the Watchtower testing code, but if we
recover from failure there we can simplify some of the callers. In particular in
NativeObject.cpp this moves the `watchPropertyValueChange` calls closer to the
`setSlot` calls.

Differential Revision: https://phabricator.services.mozilla.com/D238936

UltraBlame original commit: 5db74dbc123fbe6bf88c76276eb6e4ff7312090c