Bug 1949273: Add a lambda option for out-of-line code. r=jandem,rhunt
Adds a new subclass of OutOfLineCodeBase that takes a lambda. This is
very useful for the majority of out-of-line code, which only needs to
defer a few masm instructions to the end of the block when emitting code
and doesn't need any complicated logic.

Differential Revision: https://phabricator.services.mozilla.com/D238825

UltraBlame original commit: ea2290563c27b11e1be4e0155e5d44f44ecd8e9d