Bug 1948185: Check for an element which has just become display: contents in DocAccessible::PruneOrInsertSubtree. If so, queue style and bounds cache updates. r=morgan
Our normal computed style and bounds change notifications don't fire in this case.
Previously, we kept the bounds from before the change to display: contents in the cache.
Any new children added to the display: contents element would use their grandparent for parent relative bounds, since display: contents elements have no frame and thus can't be used for parent relative bounds.
When we accumulated bounds for the children in RemoteAccessible using their ancestors, we included the stale bounds for the element that is now display: contents.
This resulted in double counting the distance between the parent and the element before it became display: contents.

PruneOrInsertSubtree gets called for elements which are losing a frame, so we now check for this case here and update the cache appropriately.

Differential Revision: https://phabricator.services.mozilla.com/D239081

UltraBlame original commit: 13ea5f5a99123c8bb2469188beecdb2864dbd8f2