Bug 1947767 part 11 - Rely on fuses in CacheIR code. r=mgaudet
This changes the CacheIR code to match the VM code changes from the previous patches.

Telemetry shows that the `OptimizeGetIterator` fuse is popped on less than 0.05% of
websites. With the first two patches in this stack this might drop even more.

The `ArrayIteratorPrototype` fuse is a subset of the `OptimizeGetIterator` fuse
so worst-case its numbers should be similar.

Differential Revision: https://phabricator.services.mozilla.com/D238573

UltraBlame original commit: 2a0b3de2486f9fb43b5279d8856ca6dcb23d41d2