Bug 1949079 - explicitly mark maximized windows as not fullscreen  r=win-reviewers,handyman
Windows appears to detect maximized windows with custom titlebars as
"full screen", and hides the taskbar behind them when autohide is
enabled. (This appears to have been due to a change to their explicitly
undocumented heuristics.)

To avert this, arrange for just-maximized windows to be marked as _not_
being fullscreen. Per the relevant documentation this should have no
effect, but it observably does -- it doesn't fix the issue in all cases,
but it does for the common cases of "clicking the maximize icon" and
"restoring maximized windows at Firefox start".

Differential Revision: https://phabricator.services.mozilla.com/D239277

UltraBlame original commit: 9b47ac760d41d3af29abdf883d8afd85735699fd