Bug 1935894 - Pref out hidden taskbar window workarounds. r=win-reviewers,rkraesig
This seems to fix the issue according to the bug comments (see
comment 61 and comment 17).

Put this behind a pref in case it causes worse regressions. Idea is to
remove the code in a release or two.

Differential Revision: https://phabricator.services.mozilla.com/D239425

UltraBlame original commit: aa2ffbf4218c73bf7e9d7ec0d49d6968cd300ec5