Bug 1950032 - disable some wpt test items on Linux. r=jrmuizel
In these wpts, they will append both mp4 and webm sources and let browser to
determine which source is supported. The mp4 file is HEVC, which doesn't support
on Linux before until we turn the pref on in this bug.

If we run webm file on the test, the test will pass as before. Therefore, we
should either fix this color diff error for HEVC, or modify the test to use
webm file first.

Differential Revision: https://phabricator.services.mozilla.com/D239628

UltraBlame original commit: b3b65f1c13d5f1078f8394a26a75ebad6bde252d