Bug 1940773: Keep a reference to SimpleTest in AccessibilityUtils. r=ayeddi,robwu
For some reason I don't understand, window.SpecialPowers.SimpleTest and window.SimpleTest sometimes seem to be cleared during tests (or the global scope changes completely).
Ultimately, we don't control the test scope, so it's best not to rely on it anyway.
Instead, save a reference to SimpleTest in the AccessibilityUtils scope in AccessibilityUtils.init so we can always rely on it being there, since AccessibilityUtils completely controls that scope.
This patch also re-enables the test which was perma failing on CI.

Differential Revision: https://phabricator.services.mozilla.com/D239067

UltraBlame original commit: 3548a4263a84e469f91226bda7a1e87c30aff202