Bug 1947748 - Register only Gecko pings when running on geckoview r=TravisLong,chutten
Now why did this cause problems?
Fenix registers pings in Kotlin code, which calls back into Glean.
Glean stores the registered instance in an internal hashmap for later
lookup.
These instances carry around state (whether they are enabled or not),
which `setEnabled` changes.
Now eventually `Services.fog.registerCustomPings()` is called in
geckoview.js, which re-registers all pings (because it wasn't limited to
a subset based on the app ID).
This overwrites the stored instances in the Glean-internal hashmap and
thus also the state.
So later when looking up whether a ping is enabled and Glean should
store data for a ping it checks the now-latest instance and that's a
different one from the one `setEnabled` was called on in Fenix.
The `usage-reporting` ping was thus always marked as disabled in this
scenario.

Why did it work in simple testing?
Because Gecko never launched!
A quick open & background of Fenix doesn't even start Gecko
and thus never calls `registerCustomPings` and everything works.
Most users however do have tabs open and thus Gecko is launched and the
pings will lack the data.

This is hard to write a proper test for. Eventually we should have a
Fenix instrumentation test ensuring the ping is sent with valid data
after tabs have been opened.

Differential Revision: https://phabricator.services.mozilla.com/D239513

UltraBlame original commit: 4fb05c43306db47691205a3314b927a39109c230