Bug 1950596 - Avoid creating Canvas2D target until actually drawing. r=aosmond
This mainly tries to avoid referring to mTarget inside operations that are only
modifying the current transforms or clips. These can be represented in ContextState
while there is no actual target. When we do hit an actual drawing operation,
then we can instantiate the target and push the ContextState back onto it, while
also taking note if that drawing operation actually clears the target.

This way, if there are transforms or clips that precede a clear at the beginning
of a frame, we can avoid having to copy the frame contents from the previous
frame that would have normally been triggered by modifying the transforms or
clips.

Differential Revision: https://phabricator.services.mozilla.com/D239727

UltraBlame original commit: 9cc733bde4acd05b9c2f228499534a31b283d1e7