Bug 1950087 part 4 - Avoid calls to ArraySpeciesCreate in self-hosted code for plain arrays. r=mgaudet
This adds a `CanOptimizeArraySpecies` intrinsic that returns true if we don't need to call
`ArraySpeciesCreate`. This accounts for > 98% of all calls on Speedometer and JetStream.
The next patch will inline this intrinsic in JIT code.

This makes the `map` micro-benchmark in the bug about 2x faster.

Differential Revision: https://phabricator.services.mozilla.com/D239706

UltraBlame original commit: 4c60476a491f6fe756508d1cf3a7d6f17554a13e