Bug 1950751 [wpt PR 50980] - [PEPC] Fix display of PEPC in quirks mode when height: auto is used, a=testonly
Automatic update from web-platform-tests
[PEPC] Fix display of PEPC in quirks mode when height: auto is used

In quirks mode when `height: auto` is used and there is no set height on
any of the elements, the browser will try to set the height to the
entire viewport. However the permission element has special height
restrictions and special interactions between padding and height so this
results in incorrect behavior which can be abused: the element itself
will stay at the predefined max-height, but the child text node will
have a height of 100% and therefore bleed outside the permission
element.

2 root causes: 1) The child does not inherit max-height from the
permission element 2) The handling of the auto + padding case is
intended to allow the padding to basically dictate the overall height as
the result of top-padding + content height + bottom padding. But if
height is `auto` it makes the content height unnecessarily large (it
should be just the text height an no extra space). So instead if we're
in this case we should treat `height: auto` as `height: fit-content`.

Fixed: 398803201
Change-Id: Ie68d58d2658468c2ccca0950d4965d2d19d61afb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304683
Reviewed-by: Anders Hartvoll Ruud <andruudchromium.org>
Commit-Queue: Andy Paicu <andypaicuchromium.org>
Cr-Commit-Position: refs/heads/main{#1425622}

--

wpt-commits: f6b9929be02f5cf641c6ae9b265d2c6a9177ccd0
wpt-pr: 50980

UltraBlame original commit: aedb4b8c25c5b5852ba0402196b48d547013d098