Bug 1665474 - Part 11: Don't create DocGroups for cross-origin data documents, r=smaug,farre
This changes how DocGroups are created to make cross-origin data
documents share a DocGroup with their embedder.

This is required to better align DocGroups with Similar-origin Window
Agents (which correlate with DocGroups), which is required for other
parts of this patch stack.

Differential Revision: https://phabricator.services.mozilla.com/D239226

UltraBlame original commit: 2727c78f7920a0cdc2fedcffc041f8571292fed0