Bug 1949914 - Clear all cached WR resources when hiding a popup. r=gfx-reviewers,gw
It is the right thing to do, and fixes this. Otherwise we need to
somehow invalidate SWGL after showing a window...

It's been 3 years since we last tried to do it and it is early in the
cycle. If intermittents show up we can fix them. We can fix this
regression in 137 by backout.

For reference, another thing that worked, was forcing a full render
in RendererOGL.cpp, but that's obviously not ok, generally.

Differential Revision: https://phabricator.services.mozilla.com/D240397

UltraBlame original commit: 9d75dd384f425ca22bb19776e5c25647e34ba575