Bug 1951861 Part 1 - Adjust flexible track sizes in grid-max-sizing-flex-006.html to make the row sizes align with a quarter of a pixel. r=dholbert
The motivation of this patch is that this test would fail after we fix
bug 1481876, and it already fails in Chrome and Safari, which is worth fixing
now.

This patch maintains the intent of the sub-test that validates fractional `fr`
units. However, the flex factor sum, which used to be less than one, now becomes
greater than one. Therefore, I have distilled this sub-test into a web-platform
test in the next part to exercise the code path and the behavior described in
spec 12.7.1.2 [1].

[1] https://drafts.csswg.org/css-grid-2/#algo-find-fr-size

Differential Revision: https://phabricator.services.mozilla.com/D240553

UltraBlame original commit: 1ab98d96d49e52552e806c0516b901bf55315952