Bug 1948811: Compare windows instead of tabs to show canvas permission prompt. r=mconley
Each window gets the observe call once. gBrowser.selectedBrowser refers to the currently selected tab and we check against that to verify if the target browser instance is in the current window.

This patch changes it to comparing windows instead of browser instances.

Differential Revision: https://phabricator.services.mozilla.com/D238974

UltraBlame original commit: afeac0ccd8625382bb5bc077d833722330d413c1