Bug 1947470 - Rename visual-scrollIntoView.html and add another variant of the test. r=dlrobertson
A big difference of the new test is that it's testing on the root scroll
container which has been already scrolled to some extent.

Just like the original test, this new test doesn't work on Firefox
because of the lack of the support of touch actions.

Differential Revision: https://phabricator.services.mozilla.com/D239482

UltraBlame original commit: 79a9ca542d3ea8a12c38717de275fe8d3f869083